Refactor transaction mocks
