+++ packages/ui/test/modals/InviteMemberModal.test.tsx
@@ -14,7 +14,7 @@ import { UseApi } from '../../src/providers/api/provider'
 import { KeyringContext } from '../../src/providers/keyring/context'
 import { MockQueryNodeProviders } from '../helpers/providers'
 import { selectMember } from '../helpers/selectMember'
-import { stubTransaction, stubTransactionFailure, stubTransactionSuccess } from '../helpers/transactions'
+import { stubTransaction, stubTransactionFailure, stubTransactionSuccess } from '../mocks/transactions'
 import { alice, aliceStash, mockKeyring } from '../mocks/keyring'
 import { getMember } from '../mocks/members'
 import { setupMockServer } from '../mocks/server'
