+++ packages/ui/test/modals/UpdateMembershipModal.test.tsx
@@ -12,7 +12,7 @@ import { ApiContext } from '../../src/providers/api/context'
 import { UseApi } from '../../src/providers/api/provider'
 import { KeyringContext } from '../../src/providers/keyring/context'
 import { MockQueryNodeProviders } from '../helpers/providers'
-import { stubTransaction, stubTransactionFailure, stubTransactionSuccess } from '../helpers/transactions'
+import { stubTransaction, stubTransactionFailure, stubTransactionSuccess } from '../mocks/transactions'
 import { alice, bob, mockKeyring } from '../mocks/keyring'
 import { getMember, MockMember } from '../mocks/members'
 import { setupMockServer } from '../mocks/server'
